+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 **0.0.3**
 * Patch fix to getAcceptLanguage and getFromHeader. Both are independent from the accept class.
 * Updated readme
+
+**0.0.4**
+* Added support for top level domain (aka domain). 
+* Updated readme
+* Fixes the `this.locale` to be set when whatever function is called.
+* Added more tests